+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 3.0.6
 
+- Focus node added in UpButton
 - OnBlur added in UpTextfield
-- Updated UpButton Widget
 - Updated UpSecurity
 
 ## 3.0.5